app/cabin_recorder/requirements.txt
tk
pillow
tkdesigner
pywebview
pupil-labs-realtime-api